Loading in triangle verticies, no colour. (It's upside down!)

+++ Source/Triangle.cpp
@@ -1,7 +1,7 @@
 #include "Triangle.h"
 
 
-Triangle::Triangle(glm::vec3 v0,glm::vec3 v1,glm::vec3 v2){
+Triangle::Triangle(glm::ivec3 v0,glm::ivec3 v1,glm::ivec3 v2){
   vertices[0] = v0;
   vertices[1] = v1;
   vertices[2] = v2;
